+++ lib/Api/PrepaidContractsApi.php
@@ -7,14 +7,14 @@ use Secuconnect\Client\ApiException;
 use Secuconnect\Client\Authentication\Authenticator;
 
 /**
- * ServicesUploadidentsProductApi
+ * PrepaidContractsApi
  *
  * @category Class
  * @package  Secuconnect\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
-class ServicesUploadidentsProductApi
+class PrepaidContractsApi
 {
     /**
      * API Client
@@ -52,7 +52,7 @@ class ServicesUploadidentsProductApi
      *
      * @param ApiClient $apiClient set the API client
      *
-     * @return ServicesUploadidentsProductApi
+     * @return PrepaidContractsApi
      */
     public function setApiClient(ApiClient $apiClient)
     {
@@ -61,33 +61,41 @@ class ServicesUploadidentsProductApi
     }
 
     /**
-     * Operation addUploadidents
+     * Operation mappingZvt
      *
-     * POST Services/Uploadidents
+     * POST Prepaid/Contracts/me/mappingZvtmappingZvt
      *
-     * @param \Secuconnect\Client\Model\UploadidentsProductDTO $body Upload idents input properties 
+     * @param string $prepaid_contract_id Prepaid contract id (required)
+     * @param \Secuconnect\Client\Model\PrepaidZvtDTO $body Prepaid transaction input properties 
      * @throws ApiException on non-2xx response
-     * @return \Secuconnect\Client\Model\UploadidentsProductModel
+     * @return \Secuconnect\Client\Model\PrepaidMappingZvt
      */
-    public function addUploadidents($body)
+    public function mappingZvt($prepaid_contract_id, $body)
     {
-        list($response) = $this->addUploadidentsWithHttpInfo($body);
+        list($response) = $this->mappingZvtWithHttpInfo($prepaid_contract_id, $body);
         return $response;
     }
 
     /**
-     * Operation addUploadidentsWithHttpInfo
+     * Operation mappingZvtWithHttpInfo
      *
-     * POST Services/Uploadidents
+     * POST Prepaid/Contracts/me/mappingZvtmappingZvt
      *
-     * @param \Secuconnect\Client\Model\UploadidentsProductDTO $body Upload idents input properties 
+     * @param string $prepaid_contract_id Prepaid contract id (required)
+     * @param \Secuconnect\Client\Model\PrepaidZvtDTO $body Prepaid transaction input properties 
      * @throws ApiException on non-2xx response
-     * @return array of \Secuconnect\Client\Model\UploadidentsProductModel, HTTP status code, HTTP response headers (array of strings)
+     * @return array of \Secuconnect\Client\Model\PrepaidMappingZvt, HTTP status code, HTTP response headers (array of strings)
      */
-    public function addUploadidentsWithHttpInfo($body)
+    public function mappingZvtWithHttpInfo($prepaid_contract_id, $body)
     {
+        // verify the required parameter 'prepaid_contract_id' is set
+        if ($prepaid_contract_id === null || (is_array($prepaid_contract_id) && count($prepaid_contract_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $prepaid_contract_id when calling mappingZvt'
+            );
+        }
         // parse inputs
-        $resourcePath = "/Services/Uploadidents";
+        $resourcePath = "/Prepaid/Contracts/{prepaidContractId}/mappingZvt";
         $httpBody = '';
         $queryParams = [];
         $headerParams = [];
@@ -98,6 +106,14 @@ class ServicesUploadidentsProductApi
         }
         $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(['application/json']);
 
+        // path params
+        if ($prepaid_contract_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "prepaidContractId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($prepaid_contract_id),
+                $resourcePath
+            );
+        }
         // body params
         $_tempBody = null;
         if (isset($body)) {
@@ -125,15 +141,15 @@ class ServicesUploadidentsProductApi
                     $queryParams,
                     $httpBody,
                     $headerParams,
-                    '\Secuconnect\Client\Model\UploadidentsProductModel',
-                    '/Services/Uploadidents'
+                    '\Secuconnect\Client\Model\PrepaidMappingZvt',
+                    '/Prepaid/Contracts/{prepaidContractId}/mappingZvt'
                 );
 
-                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\UploadidentsProductModel', $httpHeader), $statusCode, $httpHeader];
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\PrepaidMappingZvt', $httpHeader), $statusCode, $httpHeader];
             } catch (ApiException $e) {
                 switch ($e->getCode()) {
                     case 200:
-                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\UploadidentsProductModel', $e->getResponseHeaders());
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\PrepaidMappingZvt', $e->getResponseHeaders());
                         $e->setResponseObject($data);
                         break;
                     case 401: